requirements.txt
modelscope
pyyaml
streaming_vocos
vector_quantize_pytorch
